requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
jsonpickle==0.9.5
requests==2.18.4
MarkupSafe==1.0
simplejson==3.11.1
Werkzeug==0.12.2
gunicorn==19.1.1
python-firebase==1.2
